feat(jessieOverrides): simpler .eslintrc.js API
This new function encapsulates more of the Jessie detection in
order to make .eslintrc.js files simpler.

Update README accordingly.
