+++ sample.eslintrc.js
@@ -1,19 +1,7 @@
 // eslint-disable-next-line import/no-unresolved
-const { partitionFromGlobs } = require('eslint-config-jessie/util');
+const { jessieOverrides } = require('eslint-config-jessie/util');
 
-const overrides = [];
-
-// Guard against infinite recursion.
-if (!process.env.DISABLE_JESSIE_OVERRIDE) {
-  const [jessieFiles] = partitionFromGlobs(['**/*.js'], __dirname);
-  if (jessieFiles.length) {
-    // We have some Jessie files to apply against.
-    overrides.push({
-      files: jessieFiles,
-      extends: ['jessie'],
-    });
-  }
-}
+const overrides = [...jessieOverrides(['**/*.js'], __dirname)];
 
 module.exports = {
   env: {
